+++ cider.el
@@ -161,6 +161,17 @@ command when there is no ambiguity."
   :group 'cider
   :package-version '(cider . "0.13.0"))
 
+(defcustom cider-allow-jack-in-without-project 'warn
+  "Controls what happens when doing `cider-jack-in' outside a project.
+When set to 'warn you'd prompted to confirm the command.
+When set to t `cider-jack-in' will quietly continue.
+When set to nil `cider-jack-in' will fail."
+  :type '(choice (const :tag "always" t)
+                 (const 'warn)
+                 (const :tag "never" nil))
+  :group 'cider
+  :package-version '(cider . "0.14.0"))
+
 (defcustom cider-known-endpoints nil
   "A list of connection endpoints where each endpoint is a list.
 For example: \\='((\"label\" \"host\" \"port\")).
@@ -513,12 +524,20 @@ own buffer."
                          params))
 
                (cmd (format "%s %s" command-resolved params)))
-          (when-let ((repl-buff (cider-find-reusable-repl-buffer nil project-dir)))
-            (let ((nrepl-create-client-buffer-function  #'cider-repl-create)
-                  (nrepl-use-this-as-repl-buffer repl-buff))
-              (nrepl-start-server-process
-               project-dir cmd
-               (when cljs-too #'cider-create-sibling-cljs-repl)))))
+          (if (or project-dir cider-allow-jack-in-without-project)
+              (progn
+                (when (or project-dir
+                          (eq cider-allow-jack-in-without-project t)
+                          (and (null project-dir)
+                               (eq cider-allow-jack-in-without-project 'warn)
+                               (y-or-n-p "Are you sure you want to run `cider-jack-in' without a Clojure project? ")))
+                  (when-let ((repl-buff (cider-find-reusable-repl-buffer nil project-dir)))
+                    (let ((nrepl-create-client-buffer-function  #'cider-repl-create)
+                          (nrepl-use-this-as-repl-buffer repl-buff))
+                      (nrepl-start-server-process
+                       project-dir cmd
+                       (when cljs-too #'cider-create-sibling-cljs-repl))))))
+            (user-error "`cider-jack-in' is not allowed without a Clojure project")))
       (user-error "The %s executable isn't on your `exec-path'" command))))
 
 ;;;###autoload